third_party/WebKit/LayoutTests/fast/forms/input-valueasnumber-unsupported.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Tests for .valueAsNumber with non-supported input types.');

var input = document.createElement('input');
input.type = 'range';
document.body.appendChild(input);

function checkValueAsNumber(type) {
    debug('Check for type=' + type);
    input.type = type;
    // Not allowed to set the value of file to anything except empty string.
    input.value = (type == 'file') ? '' : '0';
    shouldBe('input.valueAsNumber', 'Number.NaN');
    shouldThrow('input.valueAsNumber = 0', '"InvalidStateError: Failed to set the \'valueAsNumber\' property on \'HTMLInputElement\': This input element does not support Number values."');
}

checkValueAsNumber('button');
checkValueAsNumber('checkbox');
checkValueAsNumber('color');
checkValueAsNumber('email');
checkValueAsNumber('file');
checkValueAsNumber('hidden');
checkValueAsNumber('image');
checkValueAsNumber('khtml_isindex');
checkValueAsNumber('password');
checkValueAsNumber('radio');
checkValueAsNumber('reset');
checkValueAsNumber('search');
checkValueAsNumber('submit');
checkValueAsNumber('tel');
checkValueAsNumber('text');
checkValueAsNumber('url');
</script>
</body>
</html>
